+++ src/MindMeldModular.hpp
@@ -86,9 +86,17 @@ struct RefreshCounter {
 };
 
 
-struct Trigger : dsp::SchmittTrigger {
-	// implements a 0.1V - 1.0V SchmittTrigger (see include/dsp/digital.hpp) instead of 
-	//   calling SchmittTriggerInstance.process(math::rescale(in, 0.1f, 1.f, 0.f, 1.f))
+struct Trigger {
+	bool state = true;
+
+	void reset() {
+		state = true;
+	}
+	
+	bool isHigh() {
+		return state;
+	}
+	
 	bool process(float in) {
 		if (state) {
 			// HIGH to LOW